+++ Qi_ObjcMsgHook/Qi_ObjcMsgHook/QiLagMonitor/QiCallTraceTimeCostModel.h
@@ -22,6 +22,9 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) NSUInteger frequency;      //访问频次
 @property (nonatomic, strong) NSArray <QiCallTraceTimeCostModel *> *subCosts;
 
+@property (nonatomic, assign) uintptr_t lr;              // 自身的 Link Register
+@property (nonatomic, assign) uintptr_t callerLr;        // 调用者的 Link Register
+
 - (NSString *)des;
 
 @end
