+++ Qi_ObjcMsgHook/Qi_ObjcMsgHook/SceneDelegate.m
@@ -25,11 +25,62 @@
         self.window.rootViewController = homeNav;
         [self.window makeKeyAndVisible];
         
+        [self test1_1];
+        [self test2_1];
+        [self test3_1];
+        
+        [self test1_1];
+        [self test2_1];
+        [self test3_1];
+        
         [QiCallTrace stop];
         [QiCallTrace save];
     }
 }
 
+
+- (void)test1_1 {
+    usleep(11 * 1000);
+    [self test1_2];
+    [self test1_3];
+}
+
+- (void)test1_2 {
+    usleep(12 * 1000);
+}
+
+- (void)test1_3 {
+    usleep(13 * 1000);
+}
+
+- (void)test2_1 {
+    usleep(21 * 1000);
+    [self test2_2];
+    [self test2_3];
+}
+
+- (void)test2_2 {
+    usleep(22 * 1000);
+}
+
+- (void)test2_3 {
+    usleep(23 * 1000);
+}
+
+
+- (void)test3_1 {
+    usleep(31 * 1000);
+    [self test3_2];
+    [self test3_3];
+}
+
+- (void)test3_2 {
+    usleep(32 * 1000);
+}
+
+- (void)test3_3 {
+    usleep(33 * 1000);
+}
 
 - (void)sceneDidDisconnect:(UIScene *)scene {
     // Called as the scene is being released by the system.
